change: breakout build intersection table routes to seperate files
